Generalize training loop #51
