+++ src/components/OpenExampleProjectMenuItem/OpenExampleProjectMenuItem.tsx
@@ -13,7 +13,7 @@ import { CompileOptions } from "../../types/CompileOptions";
 import { LossFunction } from "../../types/LossFunction";
 import { Metric } from "../../types/Metric";
 import { OptimizationAlgorithm } from "../../types/OptimizationAlgorithm";
-import { getMnistModel } from "../FitClassifierDialog/FitClassifierDialog/networks";
+import { simpleCNN } from "../FitClassifierDialog/FitClassifierDialog/networks";
 import { compile } from "../../store/coroutines/classifier/compile";
 import { FitOptions } from "../../types/FitOptions";
 import { MenuItem } from "@mui/material";
@@ -168,21 +168,21 @@ export const OpenExampleProjectMenuItem = ({
     ]);
 
     //get model
-    const mnistModel = getMnistModel();
+    // const mnistModel = simpleCNN();
 
-    const compiledMnistModel = compile(mnistModel, mnistCompileOptions);
+    // const compiledMnistModel = compile(mnistModel, mnistCompileOptions);
 
     const mnistClassifier = {
-      compiled: compiledMnistModel,
+      // compiled: compiledMnistModel,
       data: [trainXs, trainYs],
       fitOptions: mnistFitOptions,
       inputShape: { r: 28, c: 28, channels: 1 },
       learningRate: mnistCompileOptions.learningRate,
       lossFunction: mnistCompileOptions.lossFunction,
       metrics: mnistCompileOptions.metrics,
-      model: mnistModel,
+      // model: mnistModel,
       modelMultiplier: "0.0",
-      modelName: "MnistCNN",
+      modelName: "SimpleCNN",
       modelVersion: "1",
       optimizationAlgorithm: mnistCompileOptions.optimizationAlgorithm,
       testPercentage: 0.2,
